+++ ChangeLog.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add an `ifm3d` specific `user-agent` for `http` request
 - Add support for firmware files larger than 2 GB
+- Add Shell (`bash`, `zsh`, `pwsh`) Generation completion for `ifm3d` CLI commands `<TAB>` completion
 
 ## 2.0.1 - 2026-03-09